requirements.txt
#Paquetes requeridos para la app:
Flask                           #1 Flask
Flask-SQLAlchemy                #2 SQLAlchemy con Flask.
mysql-connector-python          #Driver de Python para MySQL 
python-dotenv                   #Para cargar variables de config desde un archivo .env
Flask-WTF                       #Para el manejo de formularios en Flask
pytz                            #Libreria para TimeZones y así mostrar hora en Chile
gunicorn                        #Servidor WSGI de producción 